Pull SongDetails into sub-component folder and extract recommendations
