+++ src/Library/SongDetails/index.ts
@@ -1,0 +1,3 @@
+import SongDetails from './SongDetails'
+
+export default SongDetails
