+++ src/Library/graphql.ts
@@ -1,29 +1,4 @@
 import gql from 'graphql-tag'
-import { User } from 'react-feather'
-
-export type SongQuery = {
-  data: {
-    song: Song
-  }
-  vars: {
-    id: string
-  }
-}
-
-export const SONG_QUERY = gql`
-  query Song($id: ID!) {
-    song(id: $id) {
-      id
-      durationInSeconds
-      name
-      youtubeId
-      tags {
-        id
-        name
-      }
-    }
-  }
-`
 
 export type SongsQuery = {
   data: {
@@ -113,44 +88,6 @@ export const REMOVE_FROM_LIBRARY = gql`
   }
 `
 
-export type RecommendationCreate = {
-  data: {}
-  vars: {
-    youtubeId: string
-    recommendToUser: string
-  }
-}
-
-export const RECOMMENDATION_CREATE = gql`
-  mutation RecommendationCreate($youtubeId: ID!, $recommendToUser: ID!) {
-    recommendationCreate(input: { youtubeId: $youtubeId, recommendToUser: $recommendToUser }) {
-      errors
-    }
-  }
-`
-
-export type TeamQuery = {
-  data: {
-    team: {
-      users: User[]
-    }
-  }
-  vars: {
-    id: string
-  }
-}
-
-export const TEAM_QUERY = gql`
-  query Team($id: ID!) {
-    team(id: $id) {
-      users {
-        id
-        name
-      }
-    }
-  }
-`
-
 export type Song = {
   id: string
   durationInSeconds: number
@@ -161,11 +98,6 @@ export type Song = {
 }
 
 export type Tag = {
-  id: string
-  name: string
-}
-
-export type User = {
   id: string
   name: string
 }
